+++ packages/sitecore-jss-proxy/src/middleware/editing/index.test.ts
@@ -1,10 +1,24 @@
-import { expect } from 'chai';
+/* eslint-disable no-unused-expressions */
+import sinon from 'sinon';
 import express from 'express';
 import request from 'supertest';
-import { editingRouter } from './index';
+import { editingRouter, EditingRouterConfig } from './index';
+import { GraphQLRequestClient } from '@sitecore-jss/sitecore-jss';
+import { EditingRenderEndpointOptions } from './render';
 
 describe('editingRouter', () => {
-  const defaultOptions = {
+  const clientFactory = GraphQLRequestClient.createClientFactory({
+    endpoint: 'http://site/?sitecoreContextId=context-id',
+  });
+
+  const renderView: sinon.SinonStub = sinon.stub();
+
+  const renderConfig: EditingRenderEndpointOptions = {
+    clientFactory,
+    renderView,
+  };
+
+  const defaultOptions: EditingRouterConfig = {
     config: {
       components: ['component1', 'component2'],
       metadata: {
@@ -14,211 +28,68 @@ describe('editingRouter', () => {
         },
       },
     },
+    render: renderConfig,
   };
 
   let app: express.Express;
 
   beforeEach(() => {
     app = express();
+
+    process.env.JSS_EDITING_SECRET = 'correct';
   });
 
-  it('should throw 401 CORS error when requested origin is not allowed', (done) => {
+  afterEach(() => {
+    delete process.env.JSS_EDITING_SECRET;
+  });
+
+  it('should response 401 CORS error when requested origin is not allowed', (done) => {
     app.use(editingRouter(defaultOptions));
 
     request(app)
       .get('/config')
       .set('origin', 'http://not-allowed.com')
-      .expect(401)
-      .end((err, res) => {
-        if (err) return done(err);
-        expect(res.body.html).to.equal(
-          '<html><body>Requests from origin http://not-allowed.com not allowed</body></html>'
-        );
-        done();
-      });
+      .expect(401, 'Requests from origin http://not-allowed.com are not allowed', done);
   });
 
-  it('should throw 401 error when editing secret is not set', (done) => {
+  it('should response 401 error when editing secret is not set', (done) => {
+    process.env.JSS_EDITING_SECRET = '';
+
     app.use(editingRouter(defaultOptions));
 
     request(app)
       .get('/config')
-      .expect(401)
-      .end((err, res) => {
-        if (err) return done(err);
-        expect(res.body.html).to.equal(
-          '<html><body>Missing editing secret - set JSS_EDITING_SECRET environment variable</body></html>'
-        );
-        done();
-      });
+      .expect(401, 'Missing editing secret - set JSS_EDITING_SECRET environment variable', done);
   });
 
-  it('should throw 401 error when editing secret is incorrect', (done) => {
-    process.env.JSS_EDITING_SECRET = 'correct';
-
+  it('should response 401 error when editing secret is incorrect', (done) => {
     app.use(editingRouter(defaultOptions));
 
     request(app)
       .get('/config')
       .query({ secret: 'incorrect' })
-      .expect(401)
-      .end((err, res) => {
-        if (err) return done(err);
-        expect(res.body.html).to.equal('<html><body>Missing or invalid secret</body></html>');
-
-        delete process.env.JSS_EDITING_SECRET;
-
-        done();
-      });
+      .expect(401, 'Missing or invalid secret', done);
   });
 
-  it('should throw 405 error when not allowed method is used', (done) => {
-    process.env.JSS_EDITING_SECRET = 'correct';
-
+  it('should response 405 error when not allowed method is used', (done) => {
     app.use('/api/editing', editingRouter(defaultOptions));
 
     request(app)
       .post('/api/editing/config')
       .query({ secret: 'correct' })
-      .expect(405)
-      .end((err, res) => {
-        if (err) return done(err);
-        expect(res.body.html).to.equal(
-          '<html><body>Invalid request method or path POST /api/editing/config?secret=correct</body></html>'
-        );
-
-        delete process.env.JSS_EDITING_SECRET;
-
-        done();
-      });
+      .expect(405, 'Invalid request method or path POST /api/editing/config?secret=correct', done);
   });
 
-  it('should throw 405 error when not allowed path is used', (done) => {
-    process.env.JSS_EDITING_SECRET = 'correct';
-
+  it('should response 405 error when not allowed path is used', (done) => {
     app.use('/api/editing', editingRouter(defaultOptions));
 
     request(app)
       .get('/api/editing/fake-config')
       .query({ secret: 'correct' })
-      .expect(405)
-      .end((err, res) => {
-        if (err) return done(err);
-        expect(res.body.html).to.equal(
-          '<html><body>Invalid request method or path GET /api/editing/fake-config?secret=correct</body></html>'
-        );
-
-        delete process.env.JSS_EDITING_SECRET;
-
-        done();
-      });
-  });
-
-  describe('/config', () => {
-    it('should return editing config', (done) => {
-      process.env.JSS_EDITING_SECRET = 'correct';
-
-      app.use('/api/editing', editingRouter(defaultOptions));
-
-      request(app)
-        .get('/api/editing/config')
-        .query({ secret: 'correct' })
-        .expect(200)
-        .end((err, res) => {
-          if (err) return done(err);
-          expect(res.body).to.deep.equal({
-            components: ['component1', 'component2'],
-            packages: {
-              foo: '1.0.0',
-              bar: '2.0.0',
-            },
-            editMode: 'metadata',
-          });
-
-          delete process.env.JSS_EDITING_SECRET;
-
-          done();
-        });
-    });
-
-    it('should return editing config when components are a map', (done) => {
-      process.env.JSS_EDITING_SECRET = 'correct';
-
-      app.use(
-        '/api/editing',
-        editingRouter({
-          config: {
-            components: new Map([
-              ['component1', true],
-              ['component2', true],
-            ]),
-            metadata: {
-              packages: {
-                foo: '1.0.0',
-                bar: '2.0.0',
-              },
-            },
-          },
-        })
+      .expect(
+        405,
+        'Invalid request method or path GET /api/editing/fake-config?secret=correct',
+        done
       );
-
-      request(app)
-        .get('/api/editing/config')
-        .query({ secret: 'correct' })
-        .expect(200)
-        .end((err, res) => {
-          if (err) return done(err);
-          expect(res.body).to.deep.equal({
-            components: ['component1', 'component2'],
-            packages: {
-              foo: '1.0.0',
-              bar: '2.0.0',
-            },
-            editMode: 'metadata',
-          });
-
-          delete process.env.JSS_EDITING_SECRET;
-
-          done();
-        });
-    });
-
-    it('should return editing config when custom request path is set', (done) => {
-      process.env.JSS_EDITING_SECRET = 'correct';
-
-      app.use(
-        '/api/editing',
-        editingRouter({
-          config: {
-            components: ['component1'],
-            metadata: {
-              packages: {
-                foo: '1.0.0',
-              },
-            },
-            path: '/foo/config',
-          },
-        })
-      );
-
-      request(app)
-        .get('/api/editing/foo/config')
-        .query({ secret: 'correct' })
-        .expect(200)
-        .end((err, res) => {
-          if (err) return done(err);
-          expect(res.body).to.deep.equal({
-            components: ['component1'],
-            packages: {
-              foo: '1.0.0',
-            },
-            editMode: 'metadata',
-          });
-
-          delete process.env.JSS_EDITING_SECRET;
-
-          done();
-        });
-    });
   });
 });