+++ packages/sitecore-jss-proxy/src/middleware/editing/index.ts
@@ -6,6 +6,7 @@ import {
 } from '@sitecore-jss/sitecore-jss/editing';
 import { enforceCors } from '@sitecore-jss/sitecore-jss/utils';
 import { EditingConfigEndpointOptions, editingConfigMiddleware } from './config';
+import { EditingRenderEndpointOptions, editingRenderMiddleware } from './render';
 
 /**
  * Default endpoints for editing requests
@@ -26,12 +27,7 @@ export type EditingRouterConfig = {
   /**
    * Configuration for the /render endpoint
    */
-  render?: {
-    /**
-     * Custom path for the editing render endpoint
-     */
-    path?: string;
-  };
+  render: EditingRenderEndpointOptions;
 };
 
 /**
@@ -59,24 +55,21 @@ export const editingMiddleware = async (
     debug.editing(
       'invalid origin host - set allowed origins in JSS_ALLOWED_ORIGINS environment variable'
     );
-    return res.status(401).json({
-      html: `<html><body>Requests from origin ${req.headers?.origin} not allowed</body></html>`,
-    });
+    return res.status(401).send(`Requests from origin ${req.headers?.origin} are not allowed`);
   }
 
   if (!secret) {
     debug.editing('missing editing secret - set JSS_EDITING_SECRET environment variable');
 
-    return res.status(401).json({
-      html:
-        '<html><body>Missing editing secret - set JSS_EDITING_SECRET environment variable</body></html>',
-    });
+    return res
+      .status(401)
+      .send('Missing editing secret - set JSS_EDITING_SECRET environment variable');
   }
 
   if (secret !== providedSecret) {
-    debug.editing('invalid editing secret - sent "%s" expected "%s"', secret, providedSecret);
+    debug.editing('invalid editing secret - sent "%s" expected "%s"', providedSecret, secret);
 
-    return res.status(401).json({ html: '<html><body>Missing or invalid secret</body></html>' });
+    return res.status(401).send('Missing or invalid secret');
   }
 
   return next();
@@ -90,9 +83,7 @@ export const editingMiddleware = async (
 const editingNotFoundMiddleware = (req: Request, res: Response) => {
   debug.editing('invalid method or path - sent %s %s', req.method, req.originalUrl);
 
-  return res.status(405).json({
-    html: `<html><body>Invalid request method or path ${req.method} ${req.originalUrl}</body></html>`,
-  });
+  return res.status(405).send(`Invalid request method or path ${req.method} ${req.originalUrl}`);
 };
 
 /**
@@ -109,9 +100,7 @@ export const editingRouter = (options: EditingRouterConfig) => {
   router.use(editingMiddleware);
 
   router.get(options.config.path || ENDPOINTS.CONFIG, editingConfigMiddleware(options.config));
-  router.get(options.render?.path || ENDPOINTS.RENDER, () => {
-    return null;
-  });
+  router.get(options.render.path || ENDPOINTS.RENDER, editingRenderMiddleware(options.render));
 
   router.use(editingNotFoundMiddleware);