+++ packages/sitecore-jss/src/editing/index.ts
@@ -20,4 +20,5 @@ export {
   EditButtonTypes,
   mapButtonToCommand,
 } from './edit-frame';
+export { RenderMetadataQueryParams } from './models';
 export { LayoutKind } from './models';